[Step1] Move Product and ProductDao under domain package
