+++ src/test/java/kitchenpos/products/controller/ProductRestControllerTest.java
@@ -1,7 +1,7 @@
 package kitchenpos.products.controller;
 
 import kitchenpos.products.bo.ProductBo;
-import kitchenpos.products.model.Product;
+import kitchenpos.products.tobe.domain.Product;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.web.servlet.HttpEncodingAutoConfiguration;
